+++ src/events/year_2015/day_02.rs
@@ -1,5 +1,5 @@
 // TODO implement common function interface
-pub fn solve(input: &str) -> (i32, i32) {
+pub fn solve(input: &str) -> (i64, i64) {
 
     let part_1_result = part_1(input);
     let part_2_result = part_2(input);
@@ -7,35 +7,35 @@ pub fn solve(input: &str) -> (i32, i32) {
     (part_1_result, part_2_result)
 }
 
-fn part_1(input: &str) -> i32 {
-    let total_squares: i32 = input.lines().map(|line| {
+fn part_1(input: &str) -> i64 {
+    let total_squares: i64 = input.lines().map(|line| {
         let parts = line.split("x").collect::<Vec<&str>>();
 
-        let l: i32 = parts[0].parse().unwrap();
-        let w: i32 = parts[1].parse().unwrap();
-        let h: i32 = parts[2].parse().unwrap();
+        let l: i64 = parts[0].parse().unwrap();
+        let w: i64 = parts[1].parse().unwrap();
+        let h: i64 = parts[2].parse().unwrap();
 
         let sides = [l*w, w*h, h*l];
 
-        2 * sides.iter().sum::<i32>() + sides.iter().min().unwrap()
+        2 * sides.iter().sum::<i64>() + sides.iter().min().unwrap()
     }).sum();
 
     total_squares
 }
 
-fn part_2(input: &str) -> i32 {
-    let total_squares: i32 = input.lines().map(|line| {
+fn part_2(input: &str) -> i64 {
+    let total_squares: i64 = input.lines().map(|line| {
         let parts = line.split("x").collect::<Vec<&str>>();
 
-        let l: i32 = parts[0].parse().unwrap();
-        let w: i32 = parts[1].parse().unwrap();
-        let h: i32 = parts[2].parse().unwrap();
+        let l: i64 = parts[0].parse().unwrap();
+        let w: i64 = parts[1].parse().unwrap();
+        let h: i64 = parts[2].parse().unwrap();
 
         let mut sides = [l, w, h];
         sides.sort();
 
         (2 * (sides[0] + sides[1])) + (l*w*h)
-    }).sum::<i32>();
+    }).sum::<i64>();
 
     total_squares
 }
